+++ R/get_predictors.R
@@ -12,5 +12,12 @@
 #' head(get_predictors(m))
 #' @export
 get_predictors <- function(x) {
-  get_data(x)[, find_predictors(x, effects = "fixed", component = "all", flatten = TRUE), drop = FALSE]
+  dat <- get_data(x)[, find_predictors(x, effects = "fixed", component = "all", flatten = TRUE), drop = FALSE]
+
+  if (is_empty_object(dat)) {
+    print_color("Warning: Data frame is empty, probably you have an intercept-only model?\n", "red")
+    return(NULL)
+  }
+
+  dat
 }